+++ Changes.md
@@ -4,8 +4,8 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
-## v2.1.0 (2016-12-6)
-* Added info about module as argument to resolve function
+## Unreleased
+* Added name and path to resolve function
 
 ## v2.0.0 (2015-10-17)